refactor: Refactored routes
